👌 IMPROVE: Reorganise daint hybrid files
Reorganise the YAML files for the `hybrid` partition of Piz Daint so:

1. There is one setup file for each `account`, named
`setup-<account>.yaml`. For users that _do not_ want to specify the
account in the computer `prepend_text`, there is the default
`setup.yaml` file.
2. There is one `configure.yaml` file that contains sensible defaults
for most settings users do not have to adapt. The only settings the
user has to provide are `username`, `key_filename` and `proxy_jump`.
Note that `proxy_command` should not be used for daint because this
leads to connectivity issues.
3. All of the codes are account-agnostic and gathered in the `codes`
directory. The `computer` is not specified for the codes, so it will be
prompted when a user tries to set up a code, or it can be provided with
the `-Y, --computer` option.