add code registry tools
This adds a registry that allows users to place a set of yaml
files describing computers and codes installed on them in a
folder of their choice (AIIDA_CODE_REGISTRY).
When using the `load_code` or `load_computer` functions from
aiida_azq.registry, these codes are set up automatically on
demand.