Support v2.1 InstalledCode setup (#77)
From AiiDA v2.1, the `InstalledCode` is introduced for set up the code running in the computer. To instant the object, the `CodeBuilder` is not needed anymore but can simply by initialize the object from the InstalledCode class. However, the arguments are not the same as before, the `input_plugin` is change to `default_calc_job_plugin` and the `remote_abs_path` is changed to `filepath_executable`. The `on_computer` is not needed anymore. 
We add a new database for v2.1 specifically so the AWB 2.0 can use it to set up the computer/code from these database.